Moved Property system from analog to core
