+++ src/app/analog/models/eeobjects/measurement/Oscilloscope.ts
@@ -4,7 +4,11 @@ import {V, Vector} from "Vector";
 
 import {ClampedValue} from "math/ClampedValue";
 
-import {AnalogComponent, GenInitialInfo, Prop, PropInfo} from "analog/models/AnalogComponent";
+import {GenInitialInfo} from "core/utils/PropInfoUtils";
+
+import {Prop, PropInfo} from "core/models/PropInfo";
+
+import {AnalogComponent} from "analog/models/AnalogComponent";
 
 import {SidePositioner} from "analog/models/ports/positioners/SidePositioner";
