+++ src/app/analog/models/eeobjects/sources/CurrentSource.ts
@@ -4,10 +4,11 @@ import {V} from "Vector";
 
 import {ClampedValue} from "math/ClampedValue";
 
+import {GenInitialInfo} from "core/utils/PropInfoUtils";
+
 import {AnalogComponent} from "analog/models";
 
-import {GenInitialInfo} from "analog/models/AnalogComponent";
-import {AmperageInfo}   from "analog/models/Units";
+import {AmperageInfo} from "analog/models/Units";
 
 import {TopBottomPositioner} from "analog/models/ports/positioners/TopBottomPositioner";
